Refactor and enhance various components
Added noexcept specifiers to functions in VectorMath.h and ScrollBar_PG.h. Introduced setSignificantFigures template function. Replaced unique_ptr with shared_ptr in MenuHandler_PG.h and MenuHandler_PG.cpp. Updated ScrollBar_PG and View_PG functions for better parameter handling and view management. Corrected type casting in AstarAlgorithm_PG.cpp. Enhanced InventoryMenu_PG.cpp for better scrollbar interaction. Commented out debug prints and updated loot text in LandMassHandler_PG.cpp. Improved view updates in WorldDefault_PG.cpp and World_PG.cpp.

+++ PirateGame/Resource/AstarAlgorithm_PG.cpp
@@ -150,7 +150,7 @@ std::vector<sf::Vector2i> AStar::reconstructPath(std::shared_ptr<AstarNode> node
 }
 
 sf::Vector2i AStar::vectorToGrid(const sf::Vector2f& vector) const {
-    return sf::Vector2i(std::floor(vector.x / tileSize), std::floor(vector.y / static_cast<float>(tileSize)));
+    return sf::Vector2i(std::floor(vector.x / static_cast<float>(tileSize)), std::floor(vector.y / static_cast<float>(tileSize)));
 }
 
 sf::Vector2f AStar::gridToVector(const sf::Vector2i& grid) const {
